+++ src/hooks/api/goalsAPI/useGetGoals.ts
@@ -4,10 +4,10 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 const useGetGoals = (size = 20) =>
   useInfiniteQuery({
     queryKey: ['goals', size],
-    queryFn: ({ pageParam = 0 }) => goalsAPI.getGoals(pageParam, size),
+    queryFn: ({ pageParam }) => goalsAPI.getGoals(pageParam, size),
     getNextPageParam: (lastPage) =>
       lastPage.data.nextCursor !== null ? lastPage.data.nextCursor : undefined,
-    initialPageParam: 0,
+    initialPageParam: undefined,
   });
 
 export default useGetGoals;
