+++ src/hooks/api/notesAPI/useGetNotes.ts
@@ -4,10 +4,10 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 const useGetNotes = (goalId?: number, size = 20) =>
   useInfiniteQuery({
     queryKey: ['notes', goalId, size],
-    queryFn: ({ pageParam = 0 }) => notesAPI.getNotes(goalId, pageParam, size),
+    queryFn: ({ pageParam }) => notesAPI.getNotes(goalId, pageParam, size),
     getNextPageParam: (lastPage) =>
       lastPage.data.nextCursor !== null ? lastPage.data.nextCursor : undefined,
-    initialPageParam: 0,
+    initialPageParam: undefined,
   });
 
 export default useGetNotes;
