+++ src/hooks/api/todosAPI/useGetTodos.ts
@@ -4,11 +4,11 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 const useGetTodos = (goalId?: number, done?: boolean, size = 20) =>
   useInfiniteQuery({
     queryKey: ['todos', goalId, done, size],
-    queryFn: ({ pageParam = 0 }) =>
+    queryFn: ({ pageParam }) =>
       todosAPI.getTodos(goalId, done, pageParam, size),
     getNextPageParam: (lastPage) =>
       lastPage.data.nextCursor !== null ? lastPage.data.nextCursor : undefined,
-    initialPageParam: 0,
+    initialPageParam: undefined,
   });
 
 export default useGetTodos;
